backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-multipart==0.0.9
python-pptx==0.6.23
PyPDF2==3.0.1
python-docx==1.1.2
groq>=1.0.0
pydantic==2.9.2
Pillow==10.4.0
python-dotenv==1.0.1
aiofiles==24.1.0
# RAG dependencies
sentence-transformers>=2.2.0
scikit-learn>=1.3.0
numpy>=1.24.0
